ej4 resuelto

+++ Web/Ej4/Personaje.cs
@@ -3,9 +3,9 @@
     public abstract class Personaje
     {
         public string Nombre { get; set; }
-        public int Fuerza { get; set; }
-        public int Agilidad { get; set; }
-        public int Magia { get; set; }
+        public float Fuerza { get; set; }
+        public float Agilidad { get; set; }
+        public float Magia { get; set; }
 
         public abstract float CalcularDanio();
         public abstract string Atacar();
